projects/ml_experiments/pyproject.toml
# Purpose: Declare the ML experiments subproject package metadata and install surface.
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "birdsys-ml-experiments"
version = "0.1.0"
description = "BirdSys model training and evaluation workflows"
requires-python = ">=3.11,<3.12"
dependencies = [
  "duckdb>=1.1.3,<2",
  "matplotlib>=3.9,<4",
  "numpy>=1.26,<3",
  "pandas>=2.2,<3",
  "pillow>=10,<12",
  "pyarrow>=18,<20",
  "pyyaml>=6.0,<7",
  "timm>=1.0.12,<2",
  "torch>=2.5,<2.11",
  "torchvision>=0.20,<0.26",
  "tqdm>=4.66,<5",
]

[project.scripts]
birdsys-train-attributes = "birdsys.ml_experiments.train_attributes:main"
birdsys-train-attributes-cv = "birdsys.ml_experiments.train_attributes_cv:main"
birdsys-evaluate-model-b = "birdsys.ml_experiments.evaluate_model_b:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["birdsys.ml_experiments*"]
namespaces = true
